+++ CHANGELOG.md
@@ -15,6 +15,7 @@ Versioning](http://semver.org/spec/v2.0.0.html).
 - An RBAC rule with the `update` permission now properly authorizes PATCH
 requests.
 - eventd errors now include additional context for debugging.
+- Entities are now properly created using `sensuctl create`.
 
 ## [6.1] - 2020-10-05